56 - 1D
Write a program to Search the elements and print elements between those indexes.

Input Format

8
3 8 7 6 9 1 3 5
8
1

Constraints

Elements should be less than 10

Output Format

7 6 9

Sample Input 0

7
1 2 3 4 5 6 7
3
5
Sample Output 0

4
Explanation 0

key 3 is at the index of 2 and the another key 5 is at the index of 4. So need to print between these indices.
So here, we have only 4. So the output will be 4.